fix delay calculation

+++ import/src/main/java/com/google/cloud/healthcare/imaging/dicomadapter/backupuploader/DelayCalculator.java
@@ -2,7 +2,7 @@ package com.google.cloud.healthcare.imaging.dicomadapter.backupuploader;
 
 public class DelayCalculator {
 
-    private static final double DELAY_CALCULATION_BASE = 2d;
+    private static final long DELAY_CALCULATION_BASE = 2l;
     private static final long MILS_MUL = 1000l;
 
     private int attemptsAmount;
@@ -18,9 +18,9 @@ public class DelayCalculator {
     public long getExponentialDelayMillis(int attemptsLeft) {
         long delay = 0;
         if (attemptsLeft <= attemptsAmount) {
-            delay = minUploadDelay + Math.round(Math.pow(DELAY_CALCULATION_BASE, attemptsLeft) - DELAY_CALCULATION_BASE)*MILS_MUL;
+            delay = minUploadDelay + (Math.round(Math.pow((double)DELAY_CALCULATION_BASE, attemptsLeft)) - DELAY_CALCULATION_BASE)*MILS_MUL;
         } else {
-            delay = minUploadDelay + Math.round(Math.pow(DELAY_CALCULATION_BASE, attemptsAmount) - DELAY_CALCULATION_BASE)*MILS_MUL;
+            delay = minUploadDelay + (Math.round(Math.pow((double)DELAY_CALCULATION_BASE, attemptsAmount)) - DELAY_CALCULATION_BASE)*MILS_MUL;
         }
         long resultDelay = (delay >= (long)maxWaitingTimeBtwUpload)? (long)maxWaitingTimeBtwUpload : delay;
         System.out.println("resultDelay=" + resultDelay);
